style : styled component file move in styles folder
