+++ src/router/Router.tsx
@@ -3,8 +3,8 @@ import Home from '@pages/Home';
 import Login from '@pages/Login';
 import NotFound from '@pages/NotFound';
 import { URLS } from '@assets/urls';
-import MyPage from '@/pages/MyPage';
-import Order from '@/pages/Order';
+import MyPage from '@pages/MyPage';
+import Order from '@pages/Order';
 
 const Router = createBrowserRouter([
   {
