+++ src/styles/Order/OrderContainer/StyledReceivePersonContainer.tsx
@@ -1,0 +1,21 @@
+import styled from '@emotion/styled';
+
+export const StyledReceivePersonContainer = styled.div`
+  width: 100%;
+  margin-top: 10px;
+  display: flex;
+  flex-direction: column;
+  padding: 10px 0px;
+
+  div {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    align-items: center;
+    padding: 5px 10px;
+    gap: 14px;
+  }
+  .basic-label {
+    width: 80px;
+  }
+`;
