+++ src/styles/Order/OrderContainer/StyledSendPersonContainer.tsx
@@ -1,0 +1,21 @@
+import styled from '@emotion/styled';
+
+export const StyledSendPersonContainer = styled.div`
+  width: 100%;
+  margin-top: 10px;
+  display: flex;
+  flex-direction: column;
+  padding: 10px 0px;
+
+  input {
+    padding: 3px 8px;
+    margin: 5px 10px;
+    height: 30px;
+  }
+  p {
+    margin: 5px 10px;
+  }
+  .margin-left-20 {
+    margin-left: 20px;
+  }
+`;
