+++ src/types/hasError.d.ts
